lot of stuff changed. added best time to storage.
time - shows the actual elapsed time (and not the counter shown earlier).

timer starts on MOVE_EVENT (it was earlier LEVEL_START event).
timer stops on STOP_TIMER (it was earlier LEVEL_COMPLETE event).
(this is required to capture the time

UI change:
scores now go on right.
button panels at the bottom of scores.

undo on key 'U':

reset enabled when the game starts and remains enabled even if user undo's to start of the game.

minor refactoring.